+++ packages/compiler/ir-generation/src/__test__/convertErrorDefinition.test.ts
@@ -1,6 +1,5 @@
-import { ContainerType, ErrorDefinition, FernFilepath, TypeReference } from "@fern-api/api";
+import { ErrorDefinition, FernFilepath, TypeReference } from "@fern-api/api";
 import { convertErrorDefinition } from "../converters/convertErrorDefinition";
-import { parseInlineType } from "../utils/parseInlineType";
 
 describe("convertErrorDefinition", () => {
     it("reference to a type in another file", () => {
@@ -37,27 +36,5 @@ describe("convertErrorDefinition", () => {
         };
 
         expect(definition).toEqual(expectedDefinition);
-    });
-    it("parse inline nested containers", () => {
-        const dummyNamedType = "Dummy";
-        const dummyFernFilepath = FernFilepath.of("dummy");
-        const parsedTypeReference = parseInlineType({
-            type: "optional<list<" + dummyNamedType + ">>",
-            fernFilepath: dummyFernFilepath,
-            imports: {},
-        });
-        const expectedTypeReference = TypeReference.container(
-            ContainerType.optional(
-                TypeReference.container(
-                    ContainerType.list(
-                        TypeReference.named({
-                            fernFilepath: dummyFernFilepath,
-                            name: dummyNamedType,
-                        })
-                    )
-                )
-            )
-        );
-        expect(parsedTypeReference).toEqual(expectedTypeReference);
     });
 });
